Finish drive system.  Waiting on Sensors

+++ src/org/usfirst/frc/team342/RecycleRushRobot/commands/drive/DriveFowardDistance.java
@@ -8,8 +8,8 @@ import edu.wpi.first.wpilibj.command.Command;
 
 public class DriveFowardDistance extends Command {
 	DriveSystem drive;
-	private int distance;
-	private int targetdistance;
+	private double distance;
+	private double targetdistance;
 	
 	
 	public DriveFowardDistance(int target) {
